backend/pyproject.toml
[tool.black]
line-length = 79
target-version = ['py37']
include = '\.pyi?$'
extend-exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | foodgram/settings
    | tests
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.poetry]
name = "foodgram"
version = "0.1.0"
description = "foodgram"
authors = ["DRAGANmik <75855531+DRAGANmik@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
Pillow = "^8.3.1"
django-colorfield = "^0.4.2"
django-cors-headers = "^3.7.0"
reportlab = "^3.5.68"
gunicorn = "20.0.4"
psycopg2-binary = "2.8.6"
django-filter = "^2.4.0"
django-extra-fields = "^3.0.2"
django-rest-passwordreset = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest-django = "^4.4.0"
factory-boy = "^3.2.0"
black = "^21.7b0"
pep8-naming = "^0.12.0"
flake8-broken-line = "^0.3.0"
flake8-return = "^1.1.3"
flake8-isort = "^4.0.0"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
